Refactor schema definitions
Schema defeinitions was reworked so that we can easily override
or extend a single definition, depending on the specific need
for particular type of definition (image, overrides, module).

All share the same schema.yaml definition, but modifications are done
in SCHEMA_TYPE.yaml files. Both - extending and overriding is possible.

Fixes #32.
