+++ concreate/schema/common.yaml
@@ -1,0 +1,61 @@
+# This is a common Kwalify Schema for image or module descriptor.
+# See http://pykwalify.readthedocs.io/ for more information
+
+map:
+  name: {type: str, required: True}
+  version: {type: text, required: True}
+  schema_version: {type: int, required: True}
+  description: {type: text}
+  labels:
+    seq:
+      - map:
+          name: {type: str, required: True}
+          value: {type: str, required: True}
+          description: {type: str}
+  envs:
+    seq:
+      - map:
+          name: {type: str, required: True}
+          value: {type: any}
+          example: {type: any}
+          description: {type: str}
+  ports:
+    seq:
+      - map:
+          value: {type: int, required: True}
+          expose: {type: bool}
+          description: {type: str}
+  run:
+    map:
+      workdir: {type: str}
+      user: {type: text}
+      cmd:
+          seq:
+            - type: str
+      entrypoint:
+          seq:
+            - type: str
+  modules:
+    seq:
+      - map:
+          name: {type: str, required: True}
+          version: {type: str}
+  packages:
+    seq:
+      - {type: str}
+  artifacts: &RESOURCES
+    seq:
+      - map:
+          name: {type: str}
+          git:
+            map:
+              url: {type: str, required: True}
+              ref: {type: str}
+          file: {type: str, required: False}
+          url: {type: str, required: False}
+          md5: {type: str}
+          sha1: {type: str}
+          sha256: {type: str}
+          description: {type: str}
+        assert: "val['git'] is not None or val['file'] is not None or val['url] is not None"
+  dependencies: *RESOURCES
