+++ concreate/schema/image.yaml
@@ -1,0 +1,3 @@
+# To be merged with common.yaml
+map:
+  from: {type: str, required: True}
