--- concreate/schema/image_schema.yaml
@@ -1,86 +1,0 @@
-# This is a Kwalify Schema for image descriptor.
-# See http://pykwalify.readthedocs.io/ for more information
-map:
-  name:    {type: str, required: True}
-  from:    {type: str, required: True}
-  version: {type: text, required: True}
-  schema_version:  {type: int, required: True}
-  description: {type: text}
-  labels:
-    type: seq
-    sequence:
-      - type: map
-        mapping:
-          name: {type: str, required: True}
-          value: {type: str, required: True}
-          description: {type: str}
-  envs:
-    type: seq
-    sequence:
-      - type: map
-        mapping:
-          name: {type: str, required: True}
-          value: {type: any}
-          example: {type: any}
-          description: {type: str}
-  ports:
-    seq:
-      - map:
-          value: {type: int, required: True}
-          expose: {type: bool}
-          description: {type: str}
-  run:
-    map:
-      workdir: {type: str}
-      user: {type: text}
-      cmd:
-          type: seq
-          sequence:
-          - type: str
-      entrypoint:
-          type: seq
-          sequence:
-          - type: str
-  artifacts:
-    seq:
-      - type: map
-        mapping:
-          name: {type: str}
-          git: 
-              type: map
-              mapping:
-                  url: {type: str, required: True}
-                  ref: {type: str}
-          file: {type: str, required: False}
-          url: {type: str, required: False}
-          md5: {type: str}
-          sha1: {type: str}
-          sha256: {type: str}
-          description: {type: str}
-        assert: "val['git'] is not None or val['file'] is not None or val['url] is not None"
-  packages:
-    seq:
-      - {type: str}
-  modules:
-    seq:
-      - map:
-          name: {type: str, required: True}
-          version: {type: str}
-  dependencies:
-    seq:
-      - type: map
-        mapping:
-          name: {type: str}
-          git: 
-              type: map
-              mapping:
-                  url: {type: str, required: True}
-                  ref: {type: str}
-          file: {type: str, required: False}
-          url: {type: str, required: False}
-          md5: {type: str}
-          sha1: {type: str}
-          sha256: {type: str}
-          description: {type: str}
-        assert: "val['git'] is not None or val['file'] is not None or val['url] is not None"
-
