+++ requirements.txt
@@ -2,5 +2,5 @@ PyYAML>=3.11
 Jinja2>=2.8
 requests>=2.8.1
 six>=1.10.0
-pykwalify>=1.5.1
+pykwalify>=1.6.0
 colorlog>=2.10.0
